requirements.txt
torch
torchvision
opencv-python
pillow
numpy
matplotlib
albumentations
tqdm
scikit-image
scikit-learn
PyYAML
einops
torch-optimizer
torch-lr-finder
fastapi
uvicorn
python-multipart
-e .